backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
requests==2.31.0
httpx==0.25.2
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
PyPDF2==3.0.1
pypdf==3.17.0
beautifulsoup4==4.12.2
python-multipart==0.0.12
email-validator==2.1.0

